test: model-search schema
