+++ schemas/model-search/1.0/model-search.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "type": "array",
+  "items": {
+    "$ref": "/model_info.distribution/1.0/model_info.distribution.json#/definitions/ModelInfo"
+  },
+  "additionalItems": false
+}
