+++ schemas/web.config
@@ -35,6 +35,11 @@
           <action type="Rewrite" url="model_info.source/1.0/model_info.source.json" />
         </rule>
 
+        <rule name="model-search.json" enabled="true" stopProcessing="true">
+          <match url="^model-search\.json$" ignoreCase="false" />
+          <action type="Rewrite" url="model-search/1.0/model-search.json" />
+        </rule>
+
         <rule name="package.json" enabled="true" stopProcessing="true">
           <match url="^package\.json$" ignoreCase="false" />
           <action type="Rewrite" url="package/1.1.0/package.json" />
