+++ web.config
@@ -32,7 +32,7 @@
 
         <rule name="/model?q=..." stopProcessing="true">
           <match url="^model(/)?$" />
-          <action type="Rewrite" url="/script/model/model_search.php" appendQueryString="true" />
+          <action type="Rewrite" url="/script/model-search/model-search.php" appendQueryString="true" />
         </rule>
 
         <rule name="/model" stopProcessing="true">
